local_agent/requirements.txt
# Vito - Local Coding Agent Requirements

# Core dependencies
openai>=1.0.0              # OpenAI client (for local LLM compatibility)
fastapi>=0.104.0           # REST API framework
uvicorn[standard]>=0.24.0  # ASGI server
pydantic>=2.0.0            # Data validation
python-dotenv>=1.0.0       # Environment variable management

# Optional but recommended
black>=23.0.0              # Python code formatting
pytest>=7.4.0              # Testing framework
httpx>=0.25.0              # Async HTTP client

# For Google Gemini integration (OPTIONAL - not used in local-only mode)
# google-generativeai>=0.3.0

# System requirements
# - Python 3.10+
# - Qwen 2.5 Coder running locally via vLLM, Ollama, llama.cpp, or LM Studio
